upstream: move 'flb_upstream_conn' to it own declaration file
